+++ app/utils/pdf_processing.py
@@ -1,0 +1,13 @@
+from PyPDF2 import PdfReader
+from langchain.schema import Document
+
+def create_document(pdf):
+    pdf_reader = PdfReader(pdf)
+    text = "".join(page.extract_text() for page in pdf_reader.pages)
+
+    # Filter content: remove blank lines
+    text = "\n".join(line for line in text.split("\n") if line.strip())
+
+    # Create a Document instance with the full text
+    document = Document(page_content=text, metadata={})
+    return [document]
